(fix) Fixed the undefined `values.attributes` error thrown when saving a new patient (#675)
* Fixed the undefined patient attributes

* Attributes should be deleted for existing patients

+++ packages/esm-patient-registration-app/src/patient-registration/form-manager.ts
@@ -53,7 +53,7 @@ export class FormManager {
   ) => {
     const syncItem: PatientRegistration = {
       fhirPatient: FormManager.mapPatientToFhirPatient(
-        FormManager.getPatientToCreate(values, patientUuidMap, initialAddressFieldValues, []),
+        FormManager.getPatientToCreate(isNewPatient, values, patientUuidMap, initialAddressFieldValues, []),
       ),
       _patientRegistrationData: {
         isNewPatient,
@@ -101,6 +101,7 @@ export class FormManager {
     );
 
     const createdPatient = FormManager.getPatientToCreate(
+      isNewPatient,
       values,
       patientUuidMap,
       initialAddressFieldValues,
@@ -281,6 +282,7 @@ export class FormManager {
   }
 
   static getPatientToCreate(
+    isNewPatient: boolean,
     values: FormValues,
     patientUuidMap: PatientUuidMapType,
     initialAddressFieldValues: Record<string, any>,
@@ -303,7 +305,7 @@ export class FormManager {
         gender: values.gender.charAt(0),
         birthdate,
         birthdateEstimated: values.birthdateEstimated,
-        attributes: FormManager.getPatientAttributes(values, patientUuidMap),
+        attributes: FormManager.getPatientAttributes(isNewPatient, values, patientUuidMap),
         addresses: [values.address],
         ...FormManager.getPatientDeathInfo(values),
       },
@@ -335,7 +337,7 @@ export class FormManager {
     return names;
   }
 
-  static getPatientAttributes(values: FormValues, patientUuidMap: PatientUuidMapType) {
+  static getPatientAttributes(isNewPatient: boolean, values: FormValues, patientUuidMap: PatientUuidMapType) {
     const attributes: Array<AttributeValue> = [];
     if (values.attributes) {
       Object.entries(values.attributes)
@@ -346,26 +348,27 @@ export class FormManager {
             value,
           });
         });
+
+      if (!isNewPatient && values.patientUuid) {
+        Object.entries(values.attributes)
+          .filter(([, value]) => !value)
+          .forEach(async ([key]) => {
+            const attributeUuid = patientUuidMap[`attribute.${key}`];
+            await openmrsFetch(`/ws/rest/v1/person/${values.patientUuid}/attribute/${attributeUuid}`, {
+              method: 'DELETE',
+            }).catch((err) => {
+              console.error(err);
+            });
+          });
+      }
     }
+
     if (values.unidentifiedPatient) {
       attributes.push({
         // The UUID of the 'Unknown Patient' attribute-type will always be static across all implementations of OpenMRS
         attributeType: '8b56eac7-5c76-4b9c-8c6f-1deab8d3fc47',
         value: 'true',
       });
-    }
-
-    if (values.patientUuid) {
-      Object.entries(values.attributes)
-        .filter(([, value]) => !value)
-        .forEach(async ([key]) => {
-          const attributeUuid = patientUuidMap[`attribute.${key}`];
-          await openmrsFetch(`/ws/rest/v1/person/${values.patientUuid}/attribute/${attributeUuid}`, {
-            method: 'DELETE',
-          }).catch((err) => {
-            console.error(err);
-          });
-        });
     }
 
     return attributes;
